extract out project tests
